+++ docs/versionhistory.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 **2.1.0**
 
 - Added the possibility to specify more than one configuration file on the command line
+- Root component construction is now done after installing any alternate event loop policy provider
 
 **2.0.0** (2016-05-09)